- Added more DPDK-capable Ethernet controllers which makes mOS/mTCP compatible   with certain 1/10 Gbps NICs (e.g. X540T1, I211 and many, many others). Ack   to Juhyung. - Small coverity related fix.

+++ mtcp/src/api.c
@@ -394,7 +394,7 @@ mtcp_socket(mctx_t mctx, int domain, int type, int protocol)
 		return -1;
 	}
 
-	if (type == SOCK_STREAM) {
+	if (type == (int)SOCK_STREAM) {
 		type = (int)MTCP_SOCK_STREAM;
 	} else {
 		errno = EINVAL;
